+++ src/modules/webhook/webhook.module.ts
@@ -4,6 +4,7 @@ import { Webhook } from './entities/webhook.entity';
 import { WebhookDeliveryFailure } from './entities/webhook-delivery-failure.entity';
 import { Session } from '../session/entities/session.entity';
 import { WebhookService } from './webhook.service';
+import { WebhookDeliveryService } from './webhook-delivery.service';
 import { WebhookController } from './webhook.controller';
 import { WebhooksListController } from './webhooks-list.controller';
 import { EngineModule } from '../../engine/engine.module';
@@ -25,7 +26,7 @@ if (process.env.QUEUE_ENABLED === 'true') {
     ...queueModules,
   ],
   controllers: [WebhookController, WebhooksListController],
-  providers: [WebhookService],
+  providers: [WebhookService, WebhookDeliveryService],
   exports: [WebhookService],
 })
 export class WebhookModule {}
